+++ Server/src/main/java/org/openas2/cert/PKCS12CertificateFactory.java
@@ -1,305 +1,28 @@
 package org.openas2.cert;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.openas2.OpenAS2Exception;
-import org.openas2.Session;
-import org.openas2.WrappedException;
-import org.openas2.params.InvalidParameterException;
-import org.openas2.schedule.HasSchedule;
-import org.openas2.support.FileMonitorAdapter;
-import org.openas2.util.AS2Util;
-
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.security.GeneralSecurityException;
-import java.security.Key;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.PrivateKey;
-import java.security.cert.Certificate;
-import java.security.cert.X509Certificate;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
-public class PKCS12CertificateFactory extends BaseCertificateFactory implements AliasedCertificateFactory, KeyStoreCertificateFactory, StorableCertificateFactory, HasSchedule {
-    public static final String PARAM_FILENAME = "filename";
-    public static final String PARAM_PASSWORD = "password";
-    public static final String PARAM_INTERVAL = "interval";
-    private KeyStore keyStore;
+import org.openas2.OpenAS2Exception;
+import org.openas2.params.InvalidParameterException;
+import org.openas2.schedule.HasSchedule;
+import org.openas2.support.FileMonitorAdapter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ *  Supports keystores on the file system that require automatic refresh of the cached certs
+ *  when changes to the file occurs whilst OpenAS2 is running.
+ */
+public class PKCS12CertificateFactory extends X509CertificateFactory implements HasSchedule {
 
     private Logger logger = LoggerFactory.getLogger(PKCS12CertificateFactory.class);
-
-    public X509Certificate getCertificate(String alias) throws OpenAS2Exception {
-        try {
-            KeyStore ks = getKeyStore();
-            X509Certificate cert = (X509Certificate) ks.getCertificate(alias);
-
-            if (cert == null) {
-                throw new CertificateNotFoundException(null, alias);
-            }
-
-            return cert;
-        } catch (KeyStoreException kse) {
-            throw new WrappedException(kse);
-        }
-    }
-
-    public Map<String, X509Certificate> getCertificates() throws OpenAS2Exception {
-        KeyStore ks = getKeyStore();
-
-        try {
-            Map<String, X509Certificate> certs = new HashMap<String, X509Certificate>();
-            String certAlias;
-
-            Enumeration<String> e = ks.aliases();
-
-            while (e.hasMoreElements()) {
-                certAlias = e.nextElement();
-                certs.put(certAlias, (X509Certificate) ks.getCertificate(certAlias));
-            }
-
-            return certs;
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
 
     private int getRefreshInterval() throws InvalidParameterException {
         return getParameterInt(PARAM_INTERVAL, false);
     }
 
-    public String getFilename() throws InvalidParameterException {
-        return getParameter(PARAM_FILENAME, true);
-    }
-
-    public void setFilename(String filename) {
-        getParameters().put(PARAM_FILENAME, filename);
-    }
-
-    public KeyStore getKeyStore() {
-        return keyStore;
-    }
-
-    public void setKeyStore(KeyStore keyStore) {
-        this.keyStore = keyStore;
-    }
-
-    public char[] getPassword() throws InvalidParameterException {
-        return getParameter(PARAM_PASSWORD, true).toCharArray();
-    }
-
-    public void setPassword(char[] password) {
-        getParameters().put(PARAM_PASSWORD, new String(password));
-    }
-
-    public PrivateKey getPrivateKey(String alias) throws OpenAS2Exception {
-        if (alias == null) {
-            throw new OpenAS2Exception("Keystore alias cannot be found for method getPrivateKey(alias) call. Check that the x509_alias attribute is set correctly in the partnership.");
-        }
-        KeyStore ks = getKeyStore();
-        try {
-            PrivateKey key = (PrivateKey) ks.getKey(alias, getPassword());
-            if (key == null) {
-                throw new OpenAS2Exception("The private key was not found for alias. Check that the private key has been added to the keystore for the alias: " + alias);
-            }
-            return key;
-        } catch (GeneralSecurityException e) {
-            throw new OpenAS2Exception("Unexpected error occured fetching private key: " + e.getMessage(), e);
-        }
-    }
-
-    @SuppressWarnings("unused")
-    private PrivateKey getPrivateKey(X509Certificate cert) throws OpenAS2Exception {
-        KeyStore ks = getKeyStore();
-        String alias = null;
-
-        try {
-            alias = ks.getCertificateAlias(cert);
-
-            if (alias == null) {
-                throw new KeyNotFoundException(cert, "-- alias null from getCertificateAlias(cert) call. Check that the x509_alias attribute is set correctly in the partnership.");
-            }
-
-            PrivateKey key = (PrivateKey) ks.getKey(alias, getPassword());
-
-            if (key == null) {
-                throw new KeyNotFoundException(cert, "-- key null from getKey(" + alias + ") call. Check that the private key has been added to the keystore for the alias: " + alias);
-            }
-
-            return key;
-        } catch (GeneralSecurityException e) {
-            throw new KeyNotFoundException(cert, alias, e);
-        }
-    }
-
-    public void addCertificate(String alias, X509Certificate cert, boolean overwrite) throws OpenAS2Exception {
-        KeyStore ks = getKeyStore();
-
-        try {
-            if (ks.containsAlias(alias) && !overwrite) {
-                throw new CertificateExistsException(alias);
-            }
-
-            ks.setCertificateEntry(alias, cert);
-            save(getFilename(), getPassword());
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
-
-    public void addPrivateKey(String alias, Key key, String password) throws OpenAS2Exception {
-        KeyStore ks = getKeyStore();
-
-        try {
-            if (!ks.containsAlias(alias)) {
-                throw new CertificateNotFoundException(null, alias);
-            }
-
-            Certificate[] certChain = ks.getCertificateChain(alias);
-            if (certChain == null) {
-                X509Certificate x509cert = (X509Certificate) ks.getCertificate(alias);
-                if (x509cert.getSubjectX500Principal().equals(x509cert.getIssuerX500Principal())) {
-                    // Trust chain is to itself
-                    certChain = new X509Certificate[]{x509cert, x509cert};
-                    if (logger.isInfoEnabled()) {
-                        logger.info("Detected self-signed certificate and allowed import. Alias: " + alias);
-                    }
-                }
-            }
-            ks.setKeyEntry(alias, key, password.toCharArray(), certChain);
-
-            save(getFilename(), getPassword());
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
-
-    public void clearCertificates() throws OpenAS2Exception {
-        KeyStore ks = getKeyStore();
-
-        try {
-            Enumeration<String> aliases = ks.aliases();
-
-            while (aliases.hasMoreElements()) {
-                ks.deleteEntry(aliases.nextElement());
-            }
-
-            save(getFilename(), getPassword());
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
-
-    public void init(Session session, Map<String, String> options) throws OpenAS2Exception {
-        super.init(session, options);
-
-        // Override the password if it was passed as a system property
-        String pwd = System.getProperty("org.openas2.cert.Password");
-        if (pwd != null) {
-            setPassword(pwd.toCharArray());
-        }
-        try {
-            this.keyStore = AS2Util.getCryptoHelper().getKeyStore();
-        } catch (Exception e) {
-            throw new WrappedException(e);
-        }
-        load();
-    }
-
-    public void load(String filename, char[] password) throws OpenAS2Exception {
-        try {
-            FileInputStream fIn = new FileInputStream(filename);
-
-            load(fIn, password);
-
-            fIn.close();
-        } catch (IOException ioe) {
-            throw new WrappedException(ioe);
-        }
-    }
-
-    public void load(InputStream in, char[] password) throws OpenAS2Exception {
-        try {
-            KeyStore ks = getKeyStore();
-
-            synchronized (ks) {
-                ks.load(in, password);
-            }
-        } catch (IOException ioe) {
-            throw new WrappedException(ioe);
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
-
-    public void load() throws OpenAS2Exception {
-        load(getFilename(), getPassword());
-    }
-
-    public void removeCertificate(X509Certificate cert) throws OpenAS2Exception {
-        KeyStore ks = getKeyStore();
-
-        try {
-            String alias = ks.getCertificateAlias(cert);
-
-            if (alias == null) {
-                throw new CertificateNotFoundException(cert);
-            }
-
-            removeCertificate(alias);
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
-
-    public void removeCertificate(String alias) throws OpenAS2Exception {
-        KeyStore ks = getKeyStore();
-
-        try {
-            if (ks.getCertificate(alias) == null) {
-                throw new CertificateNotFoundException(null, alias);
-            }
-
-            ks.deleteEntry(alias);
-            save(getFilename(), getPassword());
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
-
-    public void save() throws OpenAS2Exception {
-        save(getFilename(), getPassword());
-    }
-
-    public void save(String filename, char[] password) throws OpenAS2Exception {
-        try {
-            FileOutputStream fOut = new FileOutputStream(filename, false);
-
-            save(fOut, password);
-
-            fOut.close();
-        } catch (IOException ioe) {
-            throw new WrappedException(ioe);
-        }
-    }
-
-    public void save(OutputStream out, char[] password) throws OpenAS2Exception {
-        try {
-            getKeyStore().store(out, password);
-        } catch (IOException ioe) {
-            throw new WrappedException(ioe);
-        } catch (GeneralSecurityException gse) {
-            throw new WrappedException(gse);
-        }
-    }
 
     @Override
     public void schedule(ScheduledExecutorService executor) throws OpenAS2Exception {
